requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
argparse==1.4.0
    # via -r requirements.in
beautifulsoup4==4.11.1
    # via bs4
bs4==0.0.1
    # via -r requirements.in
soupsieve==2.3.2.post1
    # via beautifulsoup4
yawrap==0.4.10
    # via -r requirements.in
